move script placement & fix pathway to js
So, just like in your NPM clone, you want to move the script tag to the bottom of your body. This is just to avoid lag time when loading as when the browser is rendering the page it has to stop and execute the script. For a file small like this it doesn't make a difference, but it is a good habit to develop. 

As for the pathway, you just added that it was in a folder called "javascripts", when it's not. So instead of src="javascripts/app.js" it would just be src="app.js". You'll be able to catch that when using the console on your browser.