moveBackward fix and syntax clean up
Great job! Tiny bug with your moveBackward(). And little syntax clean up on line 74. The welcome instructions are great. Also, I see you already have a habit of commenting which is awesome as it really will help you in the long run. (And as a side note, very nice adding the readme! Another great habit.)